pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
github = "https://github.com/samyuktha-ramesh/portfolio-chat"

[project]
name = "portfolio-chat"
version = "0.1.0"
description = "Tabular data retrieval augmented generation."
authors = [
    { name = "Samyuktha Ramesh" },
    { name = "Alexander Brady" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "ddgs>=9.6.0",
    "hydra-core>=1.3.2",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "pgeocode>=0.5.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "smolagents>=1.21.3",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "ruff>=0.12.10",
]

[tool.setuptools.package-data]
portfolio_rag = ["configs/*.yaml", "configs/**/*.yaml"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
portfolio-chat = "portfolio_chat.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]  # basics: style, flakes, isort, pyupgrade, bugbear
ignore = ["E501"]                    # let the formatter handle wrapping

[tool.ruff.lint.isort]
known-first-party = ["portfolio_chat"]

[tool.mypy]
python_version = "3.11"
files = ["src"]
exclude = ["\\.venv", "env", "build", "dist"]
follow_imports = "skip"  # don't analyze third-party libs
ignore_missing_imports = true
